login working
Guideline 5.1.1 - Legal - Privacy - Data Collection and Storage

The app requires users to register or log in to access features that are not account based.

Apps may not require users to enter personal information to function, except when directly relevant to the core functionality of the app or required by law. For example, an e-commerce app should let users browse store offerings and other features that are not account based before being asked to register, or a restaurant app should allow users to explore the menu before placing an order. Registration must then only be required for account-specific features, such as saving items for future reference or placing an order.

Next Steps

To resolve this issue, please revise the app to let users freely access the app's features that are not account based.